🔧 fix(.metadata): update revision in version and migration sections 🔧 fix(ios/Runner.xcodeproj/project.pbxproj): update PRODUCT_BUNDLE_IDENTIFIER 🔧 fix(ios/Runner/Info.plist): update CFBundleDisplayName 🔧 fix(lib/main.dart): update import statement for app_theme.dart 🔧 fix(lib/navigation_home_screen.dart): update import statements for custom_drawer and screens 🔧 fix(macos/Runner.xcodeproj/project.pbxproj): update PRODUCT_BUNDLE_IDENTIFIER
The changes in the `.metadata` file update the revision in the version and migration sections, ensuring that the correct revision is used.

In the `ios/Runner.xcodeproj/project.pbxproj` file, the `PRODUCT_BUNDLE_IDENTIFIER` is updated to `com.example.remember` to match the new app name.

The `ios/Runner/Info.plist` file is updated to change the `CFBundleDisplayName`

🐛 fix(AppInfo.xcconfig): update PRODUCT_BUNDLE_IDENTIFIER to com.example.remember
📦 chore(pubspec.yaml): update sdk version to '>=3.1.1 <4.0.0' and remove unused assets and dependencies

The PRODUCT_BUNDLE_IDENTIFIER in AppInfo.xcconfig is updated to com.example.remember to match the desired bundle identifier for the application. In pubspec.yaml, the sdk version is updated to '>=3.1.1 <4.0.0' to ensure compatibility with the latest Flutter version. Additionally, unused assets and dependencies are removed from the file to declutter the project.