🎉 feat(library.dart): add Library screen with grid and list view options 🐛 fix(library.dart): fix typo in constructor parameter
The Library screen is added, which displays a list of items in either a grid or list view. The screen also includes a bottom sheet for creating new items. The isGrid variable is used to toggle between the grid and list view. A typo in the constructor parameter of the Library class is fixed.